Resolves #10
Add new option `compressOutput=true|false` that allows to change the
behavior of the plugin. By default this option is disabled and the
plugin compresses the files based on the Parcel bundle object and that
match the `test` regular expression.

If `true` it compresses all files in the output directory and
subdirectories that match the `test` regular expression.